Change piece value from double to int

+++ src/chess/pieces/Knight.java
@@ -5,7 +5,7 @@ import java.util.*;
 
 public class Knight extends Piece {
     public static final String NOTATION = "N";
-    public static final double VALUE = 3;
+    public static final int VALUE = 3;
 
     public Knight(Board board, Square position, Color color) {
         super(board, position, color, VALUE, NOTATION);
