+++ src/chess/pieces/Queen.java
@@ -5,7 +5,7 @@ import java.util.*;
 
 public class Queen extends Piece {
     public static final String NOTATION = "Q";
-    public static final double VALUE = 9;
+    public static final int VALUE = 9;
 
     public Queen(Board board, Square position, Color color) {
         super(board, position, color, VALUE, NOTATION);
